address Dustin's comments
